rag-chatbot-backend/requirements.txt
setuptools>=70.0.0
wheel>=0.45.0
pip>=24.3
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
qdrant-client>=1.9.0
pydantic>=2.5.0,<3.0.0
python-dotenv>=1.0.0
fastembed>=0.7.4
onnxruntime>=1.17.0
numpy>=1.26.0
requests>=2.31.0
psycopg[binary]>=3.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]>=3.3.0
bcrypt==4.1.2
openai>=1.3.0
google-cloud-translate>=3.11.3
aiohttp>=3.9.0
httpx>=0.25.0